+++ app/src/main/java/com/huoergai/muse/ui/tv/TvDetailViewModel.kt
@@ -38,7 +38,7 @@ class TvDetailViewModel @Inject constructor(
     private val _tvReviews = MutableStateFlow<List<Review>>(emptyList())
     val tvReview: StateFlow<List<Review>> = _tvReviews
 
-    fun loadData(tvID: Int) {
+    fun loadData(tvID: Long) {
         viewModelScope.launch {
             val defer = listOf(
                 async {
